pyproject.toml
[tool.poetry]
name = "dsa-data"
version = "0.1.0"
description = "A dashboard for election and campaign finance data in the St Louis metro region"
authors = ["Tyler Schlichenmeyer <tyler.schlichenmeyer@gmail.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "3.9.1"
dash = "^1.17.0"
dash-leaflet = "^0.1.8"
pandas = "^1.1.4"
geopandas = "^0.8.1"
jupyterlab = "^2.2.9"
pyshp = "^2.1.2"
datatest = "^0.9.6"
numpy = "1.19.3"
dash-extensions = "^0.0.38"
gunicorn = "^20.0.4"
pytest-flask-sqlalchemy = "^1.0.2"
SQLAlchemy = "^1.3.22"
Flask-SQLAlchemy = "^2.4.4"
dash-bootstrap-components = "^0.11.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.2.0"
requests-mock = {extras = ["fixture"], version = "^1.8.0"}
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
mocked-sessions = "database.db.session"
mocked-engines = "database.engine"